Added stats
I added some statistics to determine the prediction accuracy of the
predictor when asked to predict cloud motion 5 minutes into the future.

Right now pixels are determined to be cloudy if they have a value
greater than a predefined constant (currently 1000). This may not work
well for other fits files.