Refactored code to use CloudMaps and CloudStates
CloudMap is an object containing a cartesian representation of the
sky (formerly known as CartesianSky). CloudState is an object
representing the dynamic state of the clouds. For now this only
includes the cloud velocity, but it could also include a spreading
rate in the future, or could become more complicated and, say, track
individual clouds' motion.