Refactored prediction into a CloudServer class
The class keeps track of recent cloud data which got posted to the
server. In order to predict, the CloudServer finds the median CloudState
from the past several observations and propagates that state on the most
recently-posted CloudMap to generate the prediction.

It's unclear yet whether scipy.optimize.minimize will work for
optimizing over the integers.