Created a video of predictions on ut111515 data
testPredClouds.py now uses the CloudServer to run predictions on an
entire directory, plotting the results in a grid which shows the true
cloud cover, the cloud cover predicted from five minutes ago, ten
minutes ago, and twenty minutes ago, and the difference images between
the predictions and the true cover. Velocities are calculated between
images about 2.5 minutes apart.

Interpolation is used to "solve" the problem that
scipy.optimize.minimize minimizes the rmse over the reals even though
pixel indices take on integer values. This is slow and hacky but so far
seems to be producing reasonable results. A better approach might be to
Fourier transform two images and use the phase shift to find the
velocity.