Introduced cloud map ids to increase performance
The current scipy.optimize.minimize interpolation hack causes scipy to
call _calcRmse (through _calcInterpolatedRmse) many times with the same
maps and directions. Instead of recomputing these each time, the results
are stored and returned back to the interpolation function without
computing them again. However, hashing the maps is very expensive (doing
so was about half of total runtime) so I put in cloud map ids which can
be used to uniquely identify a map instead.

Note that there was a bug in the hashing which I found but didn't look
into but which was fixed by moving to mapids.